move shared types from blueprint to core
