+++ src/core/blueprint/mod.rs
@@ -16,7 +16,6 @@ pub mod telemetry;
 mod timeout;
 mod union_resolver;
 mod upstream;
-mod wrapping_type;
 
 pub use auth::*;
 pub use blueprint::*;
@@ -31,7 +30,6 @@ pub use schema::*;
 pub use server::*;
 pub use timeout::GlobalTimeout;
 pub use upstream::*;
-pub use wrapping_type::Type;
 
 use crate::core::config::ConfigModule;
 use crate::core::try_fold::TryFold;
