+++ src/core/mod.rs
@@ -41,6 +41,7 @@ mod transform;
 pub mod try_fold;
 pub mod valid;
 pub mod worker;
+pub mod wrapping_type;
 
 // Re-export everything from `tailcall_macros` as `macros`
 use std::borrow::Cow;
@@ -48,7 +49,7 @@ use std::hash::Hash;
 use std::num::NonZeroU64;
 
 use async_graphql_value::ConstValue;
-pub use blueprint::Type;
+pub use directive::Directive;
 pub use errata::Errata;
 pub use error::{Error, Result};
 use http::Response;
@@ -56,6 +57,7 @@ use ir::model::IoId;
 pub use mustache::Mustache;
 pub use tailcall_macros as macros;
 pub use transform::Transform;
+pub use wrapping_type::Type;
 
 pub trait EnvIO: Send + Sync + 'static {
     fn get(&self, key: &str) -> Option<Cow<'_, str>>;
